requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
clean_talk_client==1.0.4
click==8.3.0
decorator==4.4.2
distro==1.9.0
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.104.1
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==6.11.0
instagrapi==2.1.5
instaloader==4.14.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Levenshtein==0.27.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
narwhals==2.5.0
networkx==3.5
numpy==1.26.4
openai==1.108.1
openpyxl==3.1.2
packaging==23.2
pandas==2.3.2
pillow==10.4.0
plotly==5.17.0
proglog==0.1.12
protobuf==4.25.8
pyarrow==21.0.0
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
streamlit==1.28.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
validators==0.35.0
zipp==3.23.0
